+++ ComputationalGeometryAlgorithms/Ray.h
@@ -14,7 +14,7 @@ public:
 	const Point3d& origin() { return m_origin; }
 	const Vector3f& direction() { return m_direction; }
 
-	Point3d at(double t)
+	Point3d at(float t)
 	{
 		return m_origin + (m_direction * t);
 	}
